refactor: LottoShop 을 domain 패키지로 이동시킴
