+++ src/test/java/domain/lotto/LottoEarningRateTest.java
@@ -1,10 +1,8 @@
 package domain.lotto;
 
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import service.LottoShop;
 
 import java.util.Arrays;
 import java.util.EnumMap;
