+++ src/test/java/domain/lotto/LottoTicketsTest.java
@@ -6,7 +6,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import service.LottoShop;
 
 import java.util.*;
 import java.util.stream.Collectors;
